Use run script to set permissions on boot
Grafana requires the data directory to be writeable, not just the files
inside it. If the direectory isn't writeable, Grafana will fail to
start.

To get around this in a docker environment, we'll use a run script which
will set the permissions of that directory on boot to 0777 to ensure
Grafana can write to that directory.

Fixes #24